added initialize.mp4
I added a tiny video to initialize wav2lip - this allows me to execute wav2lip in Step 1 without taking long because it's a tiny, still video. I want to do this so that when it comes to step 2, all necessary components have been downloaded and installed properly so that step 2 will be faster.