static_inference.py for use with static images
I discovered that Wav2LipHQ loses quality quickly when trying to lipsync static images, so this modified inference.py uses the original wav2lip when a static image is detected. It also runs incredibly faster on a static image than a video whereas Wav2LipHQ took the same amount of time as a video.